broke series list down into more components and dumbed each one down a bit more too.
