+++ client/src/components/Series/index.js
@@ -1,3 +1,7 @@
 export * from "./DriverCountDisplay"
+export * from "./FavoriteButton"
 export * from "./MultiCarDisplay"
 export * from "./SeriesList"
+export * from "./SeriesListBody"
+export * from "./SeriesListButton"
+export * from "./SeriesListHeader"
